[clean]ディレクトリ構成のリファクタリング
