+++ client/.env.sample
@@ -1,3 +1,6 @@
 
 # cesiumコンソールのアクセストークン https://ion.cesium.com
 VITE_CESIUM_ACCESS_TOKEN=
+
+# ライブサーバーのURL
+LIVE_SERVER_URL =
