+++ client/src/components/LaunchScreen/LaunchTimer.tsx
