+++ client/src/components/RocketStatus/RocketSvg.tsx
