+++ client/src/page/LaunchDataViewer.tsx
@@ -1,14 +1,14 @@
 import { BasicLayout } from "../components/Layout/BasicLayout";
-import { FlightVisualizer } from "../components/FlightVisual/FlightVisualizer";
-import { LivePlayer } from "../components/Live/LivePlayer";
-import CesiumComponent from "../components/Cesium/Cesium";
+import { FlightVisualizer } from "../components/RocketStatus/FlightVisualizer";
+import { LivePlayer } from "../components/LivePlayer";
+import CesiumComponent from "../components/Cesium/CesiumComponent";
 import { useEffect, useState } from "react";
 import React from "react";
-import TerminalLog from "../components/Terminal/TerminalLog";
+import TerminalLog from "../components/EventLog";
 import WeatherInfo from "../components/WeatherInfo";
-import VesselTelemetryViewer from "../components/VesselTelemetryViewer";
-import RealTimeChart from "../components/Chart/RealTimeChart";
-import LaunchScreen from "../components/Launch/LaunchScreen";
+import VesselTelemetryViewer from "../components/VesselTelemetry";
+import RealTimeChart from "../components/RealTimeChart";
+import LaunchScreen from "../components/LaunchScreen/LaunchScreen";
 import { EventRecord, FlightRecord } from "../types/flightRecordType";
 import { VesselTelemetryType } from "../types/vesselTelemetryType";
 import { RocketStatusType } from "../types/rocketStatusType";
